Finish renaming files
