+++ src/components/AudioDetails/AudioInfo.vue
@@ -58,8 +58,8 @@ import {
   SEND_DETAIL_PAGE_EVENT,
   DETAIL_PAGE_EVENTS,
 } from '~/store-modules/usage-data-analytics-types'
-import getProviderName from '~/utils/getProviderName'
-import getProviderLogo from '~/utils/getProviderLogo'
+import getProviderName from '~/utils/get-provider-name'
+import getProviderLogo from '~/utils/get-provider-logo'
 
 export default {
   name: 'AudioInfo',
